Initialization of the project with Node and Express
Create basic folder structure of the project: controller (api),
services and models because of the 3 layer architecture; loaders
to split the startup process of the node.js services.

Add '.npmrc' to ensure working with the same packages version
everywhere.

Add 'express' and 'compression' dependencies and 'standard' as
development dependency.

Add simple API endpoints to test the correct operation of the
app starting point.

Add brief description of the project at README.md