Add error handling for the Express' middlewares and routes
Use package <<http-errors>> to throw specific HTTP errors inside the
middlewares and routes.

Add simple custom error handling to the app as a Express middleware.
This custom error handling only expects HTTP errors so any other type
of error is considered a programmer bug and will crash the app.

Handle possible errors when starting the app:
	- Not existing <<.env>> file.
	- Fail on HTTP server listening to the specified port or pipe.

Add Express middleware to handle not existing URL with an HTTP
NotFoundError.

Handle loss of logs' entries due to an unexpected crash of the app.

Create two artificial errors on routes for educational purposes.